updated the profile creation page

+++ src/routes/mail/-components/account-switcher.tsx
@@ -1,5 +1,3 @@
-import * as React from "react"
-
 import { cn } from "@/lib/shad/utils"
 import {
   Select,
@@ -8,6 +6,8 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/shad/ui/select"
+import { useAuthStore } from "@/lib/hooks/auth"
+import { useState } from "react"
 
 interface AccountSwitcherProps {
   isCollapsed: boolean
@@ -22,12 +22,18 @@ export function AccountSwitcher({
   isCollapsed,
   accounts,
 }: AccountSwitcherProps) {
-  const [selectedAccount, setSelectedAccount] = React.useState<string>(
-    accounts[0].email
+  const profiles = useAuthStore((store) => store.profiles)
+  const [selectedProfile, setSelectedProfile] = useState(
+    profiles?.[0]
   )
 
+  if (!profiles) return null
+
   return (
-    <Select defaultValue={selectedAccount} onValueChange={setSelectedAccount}>
+    <Select defaultValue={selectedProfile?.email} onValueChange={(value) =>
+      setSelectedProfile(
+        profiles.find((profile) => profile.email === value)
+      )}>
       <SelectTrigger
         className={cn(
           "flex items-center gap-2 [&>span]:line-clamp-1 [&>span]:flex [&>span]:w-full [&>span]:items-center [&>span]:gap-1 [&>span]:truncate [&_svg]:h-4 [&_svg]:w-4 [&_svg]:shrink-0",
@@ -37,21 +43,27 @@ export function AccountSwitcher({
         aria-label="Select account"
       >
         <SelectValue placeholder="Select an account">
-          {accounts.find((account) => account.email === selectedAccount)?.icon}
+          {/*profiles?.find((profile) => profile.email === selectedProfile)?.icon*/}
           <span className={cn("ml-2", isCollapsed && "hidden")}>
             {
-              accounts.find((account) => account.email === selectedAccount)
-                ?.label
+              profiles.find((profile) => profile.email === selectedProfile?.email)
+                ?.firstName
             }
           </span>
         </SelectValue>
       </SelectTrigger>
       <SelectContent>
-        {accounts.map((account) => (
-          <SelectItem key={account.email} value={account.email}>
+        {profiles.map((profile) => (
+          <SelectItem key={profile.email} value={profile.email}>
             <div className="flex items-center gap-3 [&_svg]:h-4 [&_svg]:w-4 [&_svg]:shrink-0 [&_svg]:text-foreground">
-              {account.icon}
-              {account.email}
+              <svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                <title>Gmail</title>
+                <path
+                  d="M24 5.457v13.909c0 .904-.732 1.636-1.636 1.636h-3.819V11.73L12 16.64l-6.545-4.91v9.273H1.636A1.636 1.636 0 0 1 0 19.366V5.457c0-2.023 2.309-3.178 3.927-1.964L5.455 4.64 12 9.548l6.545-4.91 1.528-1.145C21.69 2.28 24 3.434 24 5.457z"
+                  fill="currentColor"
+                />
+              </svg>
+              {profile.email}
             </div>
           </SelectItem>
         ))}
